Move RatsShow to trajectory module
